requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
requests
python-dotenv
pyyaml
openenv-core[core]>=0.2.1
